Separate controller and web server
